+++ src/btree/include/interface.h
@@ -48,11 +48,7 @@ esForeach(napi_env env, napi_callback_info cbInfo);
 napi_value
 esForeachReverse(napi_env env, napi_callback_info cbInfo);
 
-napi_value
-esToMap(napi_env env, napi_callback_info cbInfo);
-
-napi_value
-esToSet(napi_env env, napi_callback_info cbInfo);
+#include <interface_conversion.h>
 
 #ifdef HAS_GTREE_NODE
   #include <intreface_iterators.h>